+++ src/GmicPixelFormat.cs
@@ -19,7 +19,7 @@ namespace GmicSharp
         /// <summary>
         /// An 8 bits-per-pixel grayscale format.
         /// </summary>
-        Gray,
+        Gray8,
 
         /// <summary>
         /// A 24 bits-per-pixel format; Each channel (red, green and blue) is allocated 8 bits.